+++ projects/iad-interface-admin/core/src/public_api.ts
@@ -1,0 +1,13 @@
+export * from './lib/shared/services/form-errors-stringify.service';
+export * from './lib/shared/services/event-manager.service';
+export * from './lib/shared/util/form-helper.service';
+export * from './lib/shared/util/request-params-builder';
+export * from './lib/shared/helpers/string.helper';
+export * from './lib/shared/iad-shared.module';
+
+export * from './lib/iad-primeng/table/iad-table-models';
+export * from './lib/iad-primeng/table/iad-table.component';
+export * from './lib/iad-primeng/table/iad-table-body.component';
+export * from './lib/iad-primeng/table/iad-scrollable-view.component';
+export * from './lib/iad-primeng/dom/iad-dom-handler';
+export * from './lib/iad-primeng/iad-primeng.module';