+++ projects/iad-interface-admin/partner/src/lib/common/dynamic-form/dynamic-form.component.ts
@@ -16,8 +16,7 @@ import { FormGroup } from '@angular/forms';
 import { FormControlService } from './form-control.service';
 import { FormGroupChild, FormGroupChildColumn, FormInputGroup } from './core/form-input-group';
 import { PrimeTemplate } from 'primeng/shared';
-import {FormErrorsStringifyService} from '../../shared/services/form-errors-stringify.service';
-import {FormHelperService} from '../../shared/util/form-helper.service';
+import {FormErrorsStringifyService, FormHelperService} from 'iad-interface-admin/core';
 
 @Component({
     selector: 'iad-dynamic-form',