+++ projects/iad-interface-admin/partner/src/lib/common/iad-common.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 
 import { TabbedViewComponent } from './tabbed-view/tabbed-view.component';
-import {IadSharedModule} from '../shared/iad-shared.module';
+import {IadSharedModule} from 'iad-interface-admin/core';
 
 @NgModule({
     imports: [IadSharedModule],