+++ projects/iad-interface-admin/partner/src/lib/common/tabbed-view/tabbed-view.component.ts
@@ -3,7 +3,7 @@ import { Component, Input, Output, EventEmitter, ContentChild, TemplateRef, OnCh
 import { DocumentListProjection } from '../../model/projection.model';
 import { Subject, Subscription } from 'rxjs';
 import {ProjectionTab} from './projection-tab.model';
-import {StringHelper} from '../../shared/helpers/string.helper';
+import {StringHelper} from 'iad-interface-admin/core';
 
 @Component({
     selector: 'iad-tabbed-view',