+++ projects/iad-interface-admin/partner/src/lib/data-table/data-table/data-table.component.ts
@@ -25,8 +25,7 @@ import { CustomizeQuery } from '../../filter-builder/action/customize-query';
 import {DataTableColumn, FILTER_TYPE, IDataTableColumn} from './data-table.model';
 import {DataTableInformationService} from '../services/data-table-information.service';
 import {CmsSetting} from '../services/settings-provider';
-import {IadTableComponent} from '../../iad-primeng/table/iad-table.component';
-import {ResizeEvent} from '../../iad-primeng/table/iad-scrollable-view.component';
+import {IadTableComponent, ResizeEvent} from 'iad-interface-admin/core';
 
 @Component({
     selector: 'iad-data-table',