+++ projects/iad-interface-admin/partner/src/lib/iad-framework.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { IADDataTableModule } from './data-table/iad-datatable.module';
 import { ProjectionFormModule } from './projection-form/projection-form.module';
 import { IADCommonModule } from './common/iad-common.module';
-import {IadSharedModule} from './shared/iad-shared.module';
+import {IadSharedModule} from 'iad-interface-admin/core';
 import {ToolbarModule} from './toolbar/toolbar.module';
 
 @NgModule({