+++ projects/iad-interface-admin/partner/src/lib/iad-settings-table/iad-settings-table.module.ts
@@ -4,7 +4,7 @@ import { SharedModule } from 'primeng/shared';
 import { IADDataTableModule } from '../data-table/iad-datatable.module';
 
 import { IADSettingsTableComponent } from './iad-settings-table.component';
-import {IadSharedModule} from '../shared/iad-shared.module';
+import {IadSharedModule} from 'iad-interface-admin/core';
 
 @NgModule({
     imports: [IadSharedModule, SharedModule, IADDataTableModule],