+++ projects/iad-interface-admin/partner/src/lib/projection-form/projection-form.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { DynamicFormModule } from '../common/dynamic-form/dynamic-form.module';
 
 import { ProjectionFormComponent } from './projection-form/projection-form.component';
-import {IadSharedModule} from '../shared/iad-shared.module';
+import {IadSharedModule} from 'iad-interface-admin/core';
 
 @NgModule({
     imports: [IadSharedModule, DynamicFormModule],