+++ projects/iad-interface-admin/src/lib/projection-form/projection-form.module.ts
@@ -4,9 +4,7 @@ import {IadCommonModule} from '../common/iad-common.module';
 import {DynamicFormModule} from '../common/dynamic-form/dynamic-form.module';
 
 import {FormViewComponent} from './form-view/form-view.component';
-import {IadSharedModule} from '../shared/iad-shared.module';
-
-
+import {IadSharedModule} from 'iad-interface-admin/core';
 
 @NgModule({
     imports: [