+++ projects/iad-interface-admin/src/lib/projection-grid/services/iad-projection-grid.service.ts
@@ -3,7 +3,7 @@ import { HttpClient, HttpResponse } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
 import {IadConfigService} from '../../config.service';
-import {RequestParamsBuilder} from '../../shared/util/request-params-builder';
+import {RequestParamsBuilder} from 'iad-interface-admin/core';
 
 type EntityArrayResponseType = HttpResponse<Array<any>>;