+++ Changelog.md
@@ -1,3 +1,10 @@
+#0.0.1.devel.6
+
+* ProjectionsHelper replaced to lib/common/helpers (available from 'interface-admin-ng')
+* Added property "enableInfiniteScroll" default to false IadScrollableViewComponent exported from 'interface-admin-ng/core'
+* Renamed IadTableComponent.askToRefresh to IadTableComponent.doRefresh
+* Added property enableInfiniteScroll default to false IadTableComponent exported from 'interface-admin-ng/core'
+
 #0.0.1.devel.5 
 
 * Removed fa4AngleDoubleDown from public_api